testproject/requirements-trunk.txt
django-extensions==0.9
django-jenkins==0.13.0
pep8==1.3.3
psycopg2==2.4.2
python-odesk==0.1.3
South==0.7.5
Sphinx==1.1.3
Werkzeug==0.8.3

http://github.com/django/django/tarball/master/
http://github.com/sergray/django-odesk/tarball/master/#egg=django_odesk-0.0.3-sergray
